requirements.txt
asgiref==3.7.2
Django==4.2.3
psycopg2-binary==2.9.6
sqlparse==0.4.4
tzdata==2023.3
django-jsonform==2.19.1
pillow==10.0.1
django-smart-selects = *